+++ src/main/java/plugily/projects/buildbattle/arena/impl/GuessTheBuildArena.java
@@ -28,6 +28,7 @@ import org.bukkit.entity.Player;
 import org.bukkit.inventory.Inventory;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
+import pl.plajerlair.commonsbox.minecraft.compat.ServerVersion;
 import pl.plajerlair.commonsbox.minecraft.compat.VersionUtils;
 import pl.plajerlair.commonsbox.minecraft.configuration.ConfigUtils;
 import pl.plajerlair.commonsbox.minecraft.item.ItemBuilder;
@@ -49,6 +50,9 @@ import plugily.projects.buildbattle.handlers.reward.Reward;
 import plugily.projects.buildbattle.menus.options.registry.particles.ParticleRefreshScheduler;
 import plugily.projects.buildbattle.menus.themevoter.BBTheme;
 import plugily.projects.buildbattle.user.User;
+import plugily.projects.buildbattle.utils.Debugger;
+import plugily.projects.buildbattle.utils.MessageUtils;
+
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.LinkedHashMap;
@@ -97,7 +101,7 @@ public class GuessTheBuildArena extends BaseArena {
         if(getPlayers().size() < getMinimumPlayers()) {
           if(getTimer() <= 0) {
             setTimer(getPlugin().getConfigPreferences().getTimer(ConfigPreferences.TimerType.LOBBY, this));
-            getPlugin().getChatManager().broadcast(this, getPlugin().getChatManager().colorMessage("In-Game.Messages.Lobby-Messages.Waiting-For-Players").replace("%MINPLAYERS%", String.valueOf(getMinimumPlayers())));
+            getPlugin().getChatManager().broadcast(this, getPlugin().getChatManager().colorMessage("In-Game.Messages.Lobby-Messages.Waiting-For-Players").replace("%MINPLAYERS%", Integer.toString(getMinimumPlayers())));
             return;
           }
         } else {
@@ -115,7 +119,7 @@ public class GuessTheBuildArena extends BaseArena {
           player.setLevel(getTimer());
         }
         if(getPlayers().size() < getMinimumPlayers()) {
-          getPlugin().getChatManager().broadcast(this, getPlugin().getChatManager().colorMessage("In-Game.Messages.Lobby-Messages.Waiting-For-Players").replace("%MINPLAYERS%", String.valueOf(getMinimumPlayers())));
+          getPlugin().getChatManager().broadcast(this, getPlugin().getChatManager().colorMessage("In-Game.Messages.Lobby-Messages.Waiting-For-Players").replace("%MINPLAYERS%", Integer.toString(getMinimumPlayers())));
           setArenaState(ArenaState.WAITING_FOR_PLAYERS);
           Bukkit.getPluginManager().callEvent(new BBGameStartEvent(this));
           setTimer(getPlugin().getConfigPreferences().getTimer(ConfigPreferences.TimerType.LOBBY, this));
@@ -148,9 +152,10 @@ public class GuessTheBuildArena extends BaseArena {
             Bukkit.getScheduler().runTaskLater(getPlugin(), () -> player.setGameMode(GameMode.SPECTATOR), 20);
           }
           Plot plot = getPlotManager().getPlot(getPlayers().get(round - 1));
-          if(plot != null && plot.getTeleportLocation() != null) {
+          org.bukkit.Location plotLoc = plot == null ? null : plot.getTeleportLocation();
+          if(plotLoc != null) {
             for(Player p : getPlayers()) {
-              p.teleport(plot.getTeleportLocation());
+              p.teleport(plotLoc);
               getPlugin().getRewardsHandler().performReward(p, Reward.RewardType.START_GAME, -1);
             }
           }
@@ -175,7 +180,7 @@ public class GuessTheBuildArena extends BaseArena {
           setTimer(getPlugin().getConfigPreferences().getTimer(ConfigPreferences.TimerType.THEME_SELECTION, this));
           break;
         }
-        if(!isThemeSet() && getTimer() <= 0 && currentBuilder != null) {
+        if(!themeSet && getTimer() <= 0 && currentBuilder != null) {
           Random r = new Random();
           String type = "EASY";
           switch(r.nextInt(2 + 1)) {
@@ -188,17 +193,19 @@ public class GuessTheBuildArena extends BaseArena {
             default:
               break;
           }
-          BBTheme theme = new BBTheme(getPlugin().getConfigPreferences().getThemes("Guess-The-Build_" + type)
-              .get(r.nextInt(getPlugin().getConfigPreferences().getThemes("Guess-The-Build_" + type).size())), BBTheme.Difficulty.valueOf(type));
-          setCurrentTheme(theme);
-          setThemeSet(true);
-          VersionUtils.sendActionBar(currentBuilder, getPlugin().getChatManager().colorMessage("In-Game.Guess-The-Build.Theme-Is-Name")
-              .replace("%THEME%", theme.getTheme()));
+          List<String> themes = getPlugin().getConfigPreferences().getThemes(BaseArena.ArenaType.GUESS_THE_BUILD.getPrefix() + "_" + type);
+          if (!themes.isEmpty()) {
+            BBTheme theme = new BBTheme(themes.get(r.nextInt(themes.size())), BBTheme.Difficulty.valueOf(type));
+            setCurrentTheme(theme);
+            setThemeSet(true);
+            VersionUtils.sendActionBar(currentBuilder, getPlugin().getChatManager().colorMessage("In-Game.Guess-The-Build.Theme-Is-Name")
+                .replace("%THEME%", theme.getTheme()));
+          }
           currentBuilder.closeInventory();
 
           String roundMessage = getPlugin().getChatManager().colorMessage("In-Game.Guess-The-Build.Current-Round")
-              .replace("%ROUND%", String.valueOf(round))
-              .replace("%MAXPLAYERS%", String.valueOf(getPlayers().size()));
+              .replace("%ROUND%", Integer.toString(round))
+              .replace("%MAXPLAYERS%", Integer.toString(getPlayers().size()));
           for(Player p : getPlayers()) {
         	  if (p != currentBuilder) {
         		  VersionUtils.sendTitle(p, getPlugin().getChatManager().colorMessage("In-Game.Guess-The-Build.Start-Guessing-Title"), 5, 25, 5);
@@ -208,26 +215,26 @@ public class GuessTheBuildArena extends BaseArena {
           setTimer(getPlugin().getConfigPreferences().getTimer(ConfigPreferences.TimerType.BUILD, this));
           break;
         }
-        if(getTimer() <= 90 && getCurrentTheme() != null) {
+        if(getTimer() <= 90 && currentTheme != null) {
           if(getTimer() == 90) {
             getPlugin().getChatManager().broadcast(this, getPlugin().getChatManager().colorMessage("In-Game.Guess-The-Build.Theme-Is-Long")
-                .replace("%NUM%", String.valueOf(getCurrentTheme().getTheme().length())));
+                .replace("%NUM%", Integer.toString(currentTheme.getTheme().length())));
           }
           for(Player player : getPlayers()) {
-            if(getCurrentBuilder() == player) {
+            if(currentBuilder == player) {
               continue;
             }
-            if(getWhoGuessed().contains(player)) {
-              VersionUtils.sendActionBar(player, getCurrentTheme().getTheme());
+            if(whoGuessed.contains(player)) {
+              VersionUtils.sendActionBar(player, currentTheme.getTheme());
             }
             StringBuilder actionbar = new StringBuilder();
-            for(int i = 0; i < getCurrentTheme().getTheme().length(); i++) {
-              if(Character.isWhitespace(getCurrentTheme().getTheme().charAt(i))) {
+            for(int i = 0; i < currentTheme.getTheme().length(); i++) {
+              if(Character.isWhitespace(currentTheme.getTheme().charAt(i))) {
                 actionbar.append("  ");
                 continue;
               }
-              if((getTimer() <= 75 && i == 0) || (getTimer() <= 40 && i == getCurrentTheme().getTheme().length() - 1) || (getTimer() <= 20 && i == 2) || (getTimer() <= 10 && i == 5)) {
-                actionbar.append(getCurrentTheme().getTheme().charAt(i)).append(' ');
+              if((getTimer() <= 75 && i == 0) || (getTimer() <= 40 && i == currentTheme.getTheme().length() - 1) || (getTimer() <= 20 && i == 2) || (getTimer() <= 10 && i == 5)) {
+                actionbar.append(currentTheme.getTheme().charAt(i)).append(' ');
                 continue;
               }
               actionbar.append("_ ");
@@ -235,11 +242,11 @@ public class GuessTheBuildArena extends BaseArena {
             VersionUtils.sendActionBar(player, ChatColor.AQUA + actionbar.toString());
           }
         }
-        if(getTimer() <= 0 && isThemeSet()) {
-          getPlugin().getChatManager().broadcast(this, getPlugin().getChatManager().colorMessage("In-Game.Guess-The-Build.Theme-Was-Name").replace("%THEME%", getCurrentTheme().getTheme()));
+        if(getTimer() <= 0 && themeSet) {
+          getPlugin().getChatManager().broadcast(this, getPlugin().getChatManager().colorMessage("In-Game.Guess-The-Build.Theme-Was-Name").replace("%THEME%", currentTheme.getTheme()));
           for(Player p : getPlayers()) {
             VersionUtils.sendTitles(p, getPlugin().getChatManager().colorMessage("In-Game.Guess-The-Build.Theme-Was-Title"), getPlugin().getChatManager().colorMessage("In-Game.Guess-The-Build.Theme-Was-Subtitle")
-                .replace("%THEME%", getCurrentTheme().getTheme()), 5, 25, 5);
+                .replace("%THEME%", currentTheme.getTheme()), 5, 25, 5);
           }
           currentBuilder = null;
           setThemeSet(false);
@@ -270,8 +277,9 @@ public class GuessTheBuildArena extends BaseArena {
             Plot plot = getPlotManager().getPlots().get(0);
             for(Player p : getPlayers()) {
               if (plot != null) {
-                if(plot.getTeleportLocation() != null) {
-                  p.teleport(plot.getTeleportLocation());
+                org.bukkit.Location plotLoc = plot.getTeleportLocation();
+                if(plotLoc != null) {
+                  p.teleport(plotLoc);
                 }
                 p.setPlayerWeather(plot.getWeatherType());
                 p.setPlayerTime(Plot.Time.format(plot.getTime(), p.getWorld().getTime()), false);
@@ -281,7 +289,7 @@ public class GuessTheBuildArena extends BaseArena {
             }
             currentBuilder.setGameMode(GameMode.CREATIVE);
             setTimer(getPlugin().getConfigPreferences().getTimer(ConfigPreferences.TimerType.THEME_SELECTION, this));
-            if(getArenaState() != ArenaState.IN_GAME || isThemeSet()) {
+            if(getArenaState() != ArenaState.IN_GAME || themeSet) {
               return;
             }
             for(Player player : getPlayers()) {
@@ -299,7 +307,7 @@ public class GuessTheBuildArena extends BaseArena {
           Bukkit.getPluginManager().callEvent(new BBGameEndEvent(this));
           setTimer(15);
         }
-        if(isThemeSet() && (getTimer() == (4 * 60) || getTimer() == (3 * 60) || getTimer() == 5 * 60 || getTimer() == 30 || getTimer() == 2 * 60 || getTimer() == 60 || getTimer() == 15)) {
+        if(themeSet && (getTimer() == (4 * 60) || getTimer() == (3 * 60) || getTimer() == 5 * 60 || getTimer() == 30 || getTimer() == 2 * 60 || getTimer() == 60 || getTimer() == 15)) {
           sendBuildLeftTimeMessage();
         }
         //if player leaves during round force next round
@@ -310,10 +318,10 @@ public class GuessTheBuildArena extends BaseArena {
           if(getOption(ArenaOption.IN_PLOT_CHECKER) == 1) {
             setOptionValue(ArenaOption.IN_PLOT_CHECKER, 0);
             for(Player player : getPlayers()) {
-              User builderUser = getPlugin().getUserManager().getUser(currentBuilder);
-              Plot buildPlot = builderUser.getCurrentPlot();
-              if(buildPlot != null && !buildPlot.getCuboid().isInWithMarge(player.getLocation(), 5) && buildPlot.getTeleportLocation() != null) {
-                player.teleport(buildPlot.getTeleportLocation());
+              Plot buildPlot = getPlugin().getUserManager().getUser(currentBuilder).getCurrentPlot();
+              org.bukkit.Location plotLoc = buildPlot == null ? null : buildPlot.getTeleportLocation();
+              if(plotLoc != null && !buildPlot.getCuboid().isInWithMarge(player.getLocation(), 5)) {
+                player.teleport(plotLoc);
                 player.sendMessage(getPlugin().getChatManager().getPrefix() + getPlugin().getChatManager().colorMessage("In-Game.Messages.Cant-Fly-Outside-Plot"));
               }
             }
@@ -335,7 +343,7 @@ public class GuessTheBuildArena extends BaseArena {
           scoreboardManager.stopAllScoreboards();
           teleportAllToEndLocation();
           for(Player player : getPlayers()) {
-            if(getPlugin().getConfigPreferences().getOption(ConfigPreferences.Option.BOSSBAR_ENABLED)) {
+            if(ServerVersion.Version.isCurrentEqualOrHigher(ServerVersion.Version.v1_9_R1) && getPlugin().getConfigPreferences().getOption(ConfigPreferences.Option.BOSSBAR_ENABLED)) {
               getGameBar().removePlayer(player);
             }
             player.getInventory().clear();
@@ -399,30 +407,32 @@ public class GuessTheBuildArena extends BaseArena {
 
     Random r = new Random();
 
-    List<String> themes = getPlugin().getConfigPreferences().getThemes("Guess-The-Build_EASY");
+    List<String> themes = getPlugin().getConfigPreferences().getThemes(BaseArena.ArenaType.GUESS_THE_BUILD.getPrefix() + "_EASY");
+    String itemLore = getPlugin().getChatManager().colorMessage("Menus.Guess-The-Build-Theme-Selector.Theme-Item-Lore");
+    String themeItemName = getPlugin().getChatManager().colorMessage("Menus.Guess-The-Build-Theme-Selector.Theme-Item-Name");
 
     Inventory inv = ComplementAccessor.getComplement().createInventory(null, 27, getPlugin().getChatManager().colorMessage("Menus.Guess-The-Build-Theme-Selector.Inventory-Name"));
-    inv.setItem(11, new ItemBuilder(Material.PAPER).name(getPlugin().getChatManager().colorMessage("Menus.Guess-The-Build-Theme-Selector.Theme-Item-Name")
+    inv.setItem(11, new ItemBuilder(Material.PAPER).name(themeItemName
         .replace("%theme%", !themes.isEmpty() ? themes.get(r.nextInt(themes.size())) : ""))
-        .lore(getPlugin().getChatManager().colorMessage("Menus.Guess-The-Build-Theme-Selector.Theme-Item-Lore")
+        .lore(itemLore
             .replace("%difficulty%", getPlugin().getChatManager().colorMessage("Menus.Guess-The-Build-Theme-Selector.Difficulties.Easy"))
-            .replace("%points%", String.valueOf(1)).split(";")).build());
+            .replace("%points%", Integer.toString(1)).split(";")).build());
 
-    themes = getPlugin().getConfigPreferences().getThemes("Guess-The-Build_MEDIUM");
+    themes = getPlugin().getConfigPreferences().getThemes(BaseArena.ArenaType.GUESS_THE_BUILD.getPrefix() + "_MEDIUM");
 
-    inv.setItem(13, new ItemBuilder(Material.PAPER).name(getPlugin().getChatManager().colorMessage("Menus.Guess-The-Build-Theme-Selector.Theme-Item-Name")
+    inv.setItem(13, new ItemBuilder(Material.PAPER).name(themeItemName
         .replace("%theme%", !themes.isEmpty() ? themes.get(r.nextInt(themes.size())) : ""))
-        .lore(getPlugin().getChatManager().colorMessage("Menus.Guess-The-Build-Theme-Selector.Theme-Item-Lore")
+        .lore(itemLore
             .replace("%difficulty%", getPlugin().getChatManager().colorMessage("Menus.Guess-The-Build-Theme-Selector.Difficulties.Medium"))
-            .replace("%points%", String.valueOf(2)).split(";")).build());
+            .replace("%points%", Integer.toString(2)).split(";")).build());
 
-    themes = getPlugin().getConfigPreferences().getThemes("Guess-The-Build_HARD");
+    themes = getPlugin().getConfigPreferences().getThemes(BaseArena.ArenaType.GUESS_THE_BUILD.getPrefix() + "_HARD");
 
-    inv.setItem(15, new ItemBuilder(Material.PAPER).name(getPlugin().getChatManager().colorMessage("Menus.Guess-The-Build-Theme-Selector.Theme-Item-Name")
+    inv.setItem(15, new ItemBuilder(Material.PAPER).name(themeItemName
         .replace("%theme%", !themes.isEmpty() ? themes.get(r.nextInt(themes.size())) : ""))
-        .lore(getPlugin().getChatManager().colorMessage("Menus.Guess-The-Build-Theme-Selector.Theme-Item-Lore")
+        .lore(itemLore
             .replace("%difficulty%", getPlugin().getChatManager().colorMessage("Menus.Guess-The-Build-Theme-Selector.Difficulties.Hard"))
-            .replace("%points%", String.valueOf(3)).split(";")).build());
+            .replace("%points%", Integer.toString(3)).split(";")).build());
     currentBuilder.openInventory(inv);
   }
 
@@ -435,12 +445,12 @@ public class GuessTheBuildArena extends BaseArena {
 
   @Override
   public void giveRewards() {
-    List<Map.Entry<Player, Integer>> list = new ArrayList<>(playersPoints.entrySet());
+    List<Player> list = new ArrayList<>(playersPoints.keySet());
     for(int i = 0; i <= list.size(); i++) {
       if(list.size() - 1 < i) {
         continue;
       }
-      getPlugin().getRewardsHandler().performReward(list.get(i).getKey(), Reward.RewardType.PLACE, i + 1);
+      getPlugin().getRewardsHandler().performReward(list.get(i), Reward.RewardType.PLACE, i + 1);
     }
     getPlugin().getRewardsHandler().performReward(this, Reward.RewardType.END_GAME);
   }
@@ -455,10 +465,10 @@ public class GuessTheBuildArena extends BaseArena {
         getGameBar().setTitle(getPlugin().getChatManager().colorMessage("Bossbar.Waiting-For-Players"));
         break;
       case STARTING:
-        getGameBar().setTitle(getPlugin().getChatManager().colorMessage("Bossbar.Starting-In").replace("%time%", String.valueOf(getTimer())));
+        getGameBar().setTitle(getPlugin().getChatManager().colorMessage("Bossbar.Starting-In").replace("%time%", Integer.toString(getTimer())));
         break;
       case IN_GAME:
-        getGameBar().setTitle(getPlugin().getChatManager().colorMessage("Bossbar.Time-Left").replace("%time%", String.valueOf(getTimer())));
+        getGameBar().setTitle(getPlugin().getChatManager().colorMessage("Bossbar.Time-Left").replace("%time%", Integer.toString(getTimer())));
         break;
       default:
         break;